videoFile Upload

+++ src/controllers/videoController.js
@@ -1,3 +1,4 @@
+import User from "../models/user";
 import Video from "../models/video"
 
 //Video.find({}, (error, videos)=> {});
@@ -7,27 +8,51 @@ export const home = async(req,res) => {
 }
 export const watch = async (req,res) => {
     const { id } = req.params;
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("owner");
     if(!video){
         return res.render("404", { pageTitle: "video not found"});
     }   
     return res.render("watch",{pageTitle: video.title,video});
 }
 export const getEdit = async (req,res) => {
-    const { id } = req.params;
+    const {
+        session:{
+            user:{
+                _id
+            }
+        },
+        params:{
+            id
+        }
+    } = req;
     const video = await Video.findById(id);
     if(!video){
         return res.status(404).render("404", { pageTitle: "video not found"});
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
     }   
     return res.render("edit",{ pageTitle: `Edit:${video.title}`, video});
 }
 export const postEdit = async (req,res) => {
-    const { id } = req.params;
+    const {
+        session:{
+            user:{
+                _id
+            }
+        },
+        params:{
+            id
+        }
+    } = req;
     const {title, description, hashtags} = req.body;
-    const video = await Video.exists({_id: id});
+    const video = await Video.findById(id);
     if(!video) {
         return res.render("404", {pageTitle: "video not found"});
     }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
+    }   
     await Video.findByIdAndUpdate(id, {
         title,
         description,
@@ -40,13 +65,30 @@ export const getUpload = (req,res) => {
     return res.render("upload",{pageTitle:"Upload Video"});
 }
 export const postUpload = async (req,res) => {
-    const {title, description, hashtags } = req.body;
+    const {
+        session:{
+            user:{_id}
+        },
+        body:{
+            title,
+            description, 
+            hashtags
+        },
+        file:{
+            path:fileUrl
+        }
+    } =req;
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
+            fileUrl,
+            owner:_id,
             hashtags : Video.formatHashtags(hashtags),
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
         return res.redirect("/");
     } catch(error) {
         console.log(error);
@@ -55,8 +97,27 @@ export const postUpload = async (req,res) => {
 }
 
 export const deleteVideo = async(req,res) => {
-    const {id} = req.params;
+    const {
+        session:{
+            user:{
+                _id
+            }
+        },
+        params:{
+            id
+        }
+    } = req;
+    const video = await Video.findById(id);
+    const user = await User.findById(_id)
+    if(!video) {
+        return res.render("404", {pageTitle: "video not found"});
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
+    user.videos.splice(user.videos.indexOf(id), 1);
+    user.save();   
     return res.redirect("/");
 };
